update window option

+++ src/transformer_output.py
@@ -39,40 +39,41 @@ class early_stopping():
 
 
 
-def process_one_batch(src, tgt, src_coord, tgt_coord, src_ts, tgt_ts):
-    dec_inp = torch.zeros([tgt.shape[0], patch_size**3, tgt.shape[-1]]).float().to(device)
-    dec_inp = torch.cat([tgt[:,:(tgt.shape[1]-patch_size**3),:], dec_inp], dim=1).float().to(device)
+def process_one_batch(src, tgt, src_coord, tgt_coord, src_ts, tgt_ts, patch_size):
+    x1, x2, x3 = patch_size
+    dec_inp = torch.zeros([tgt.shape[0], x1*x2*x3, tgt.shape[-1]]).float().to(device)
+    dec_inp = torch.cat([tgt[:,:(tgt.shape[1]-x1*x2*x3),:], dec_inp], dim=1).float().to(device)
     outputs = model(src, dec_inp, src_coord, tgt_coord, src_ts, tgt_ts)
 
     return outputs, tgt
 
 
-def evaluate(model,data_loader,criterion, input_type='patch', patch_size=2):
-    model.eval()      
+def evaluate(model,data_loader,criterion, patch_size):
+    model.eval()    
+    test_rollout = torch.Tensor(0)   
+    test_result = torch.Tensor(0)  
+    truth = torch.Tensor(0)
     total_loss = 0.
+    x1, x2, x3 = patch_size
     device = "cpu"
     if torch.cuda.is_available():
         device = "cuda:0"
         if torch.cuda.device_count() > 1:
             model = nn.DataParallel(model)
-    
-    with torch.no_grad():
-        if input_type == 'patch':
-            patch_size = patch_size
-        else:
-            patch_size = 1
-            
+    with torch.no_grad():        
         for i, ((src, tgt), (src_coord, tgt_coord), (src_ts, tgt_ts)) in enumerate(data_loader):
             src, tgt, src_coord, tgt_coord, src_ts, tgt_ts = src.to(device), tgt.to(device), \
                                                             src_coord.to(device), tgt_coord.to(device), src_ts.to(device), tgt_ts.to(device)
-            output, _ = process_one_batch(src, tgt, src_coord, tgt_coord, src_ts, tgt_ts)
-            total_loss += criterion(output[:,-patch_size**3:,:], tgt[:,-patch_size**3:,:]).detach().cpu().numpy()
+            dec_inp = torch.zeros([tgt.shape[0], x1*x2*x3, tgt.shape[-1]]).float().to(device)
+            dec_inp = torch.cat([tgt[:,:(tgt.shape[1]-x1*x2*x3),:], dec_inp], dim=1).float().to(device)
+            output = model(src, dec_inp, src_coord, tgt_coord, src_ts, tgt_ts)
+            total_loss += criterion(output[:,-x1*x2*x3:,:], tgt[:,-x1*x2*x3:,:]).detach().cpu().numpy()
 
     return total_loss
 
 
-def predict_model(model, test_loader, epoch, input_type='patch', patch_size=2, \
-                        plot=True, plot_range=[0,0.01], final_prediction=False):
+def predict_model(model, test_loader, epoch, patch_size, scale,\
+                    plot=True, plot_range=[0,0.01], final_prediction=False):
     '''
     plot_range: [a,b], 0<=a<b<=1, where a is the proportion that determines the start point to plot, b determines the end point. 
     final_prediction: True, if done with training and using the trained/saved model to predict the final result
@@ -83,6 +84,7 @@ def predict_model(model, test_loader, epoch, input_type='patch', patch_size=2, \
     truth = torch.Tensor(0) 
     test_ts = torch.Tensor(0)
     test_coord = torch.Tensor(0)
+    x1, x2, x3 = patch_size
     device = "cpu"
     if torch.cuda.is_available():
         device = "cuda:0"
@@ -90,11 +92,6 @@ def predict_model(model, test_loader, epoch, input_type='patch', patch_size=2, \
             model = nn.DataParallel(model)
     
     with torch.no_grad():
-        if input_type == 'patch':
-            patch_size = patch_size
-        else:
-            patch_size = 1
-
         for i, ((src, tgt), (src_coord, tgt_coord), (src_ts, tgt_ts)) in enumerate(test_loader):
             if i == 0:
                 enc_in = src
@@ -102,19 +99,19 @@ def predict_model(model, test_loader, epoch, input_type='patch', patch_size=2, \
                 test_rollout = tgt
             else:
                 enc_in = test_rollout[:,-tgt.shape[1]:,:]
-                dec_in = torch.zeros([enc_in.shape[0], patch_size**3, enc_in.shape[-1]]).float()
-                dec_in = torch.cat([enc_in[:,:(tgt.shape[1]-patch_size**3),:], dec_in], dim=1).float()
+                dec_in = torch.zeros([enc_in.shape[0], x1*x2*x3, enc_in.shape[-1]]).float()
+                dec_in = torch.cat([enc_in[:,:(tgt.shape[1]-x1*x2*x3),:], dec_in], dim=1).float()
                 #dec_in = enc_in[:,:(window_size-1),:]
             enc_in, dec_in, tgt = enc_in.to(device), dec_in.to(device), tgt.to(device)
             src_coord, tgt_coord, src_ts, tgt_ts = src_coord.to(device), tgt_coord.to(device), src_ts.to(device), tgt_ts.to(device)
             
             output = model(enc_in, dec_in, src_coord, tgt_coord, src_ts, tgt_ts)
-            test_rollout = torch.cat([test_rollout,output[:,-patch_size**3:,:].detach().cpu()],dim = 1)
+            test_rollout = torch.cat([test_rollout,output[:,-x1*x2*x3:,:].detach().cpu()],dim = 1)
 
-            test_ts = torch.cat((test_ts, tgt_ts[:,-patch_size**3:,:].flatten().detach().cpu()), 0)
-            test_coord = torch.cat((test_coord, tgt_coord[:,-patch_size**3:,:].reshape(-1,3).detach().cpu()), 0)
-            truth = torch.cat((truth, tgt[:,-patch_size**3:,:].flatten().detach().cpu()), 0)
-            test_result = torch.cat((test_result, output[:,-patch_size**3:,:].flatten().detach().cpu()), 0)
+            test_ts = torch.cat((test_ts, tgt_ts[:,-x1*x2*x3:,:].flatten().detach().cpu()), 0)
+            test_coord = torch.cat((test_coord, tgt_coord[:,-x1*x2*x3:,:].reshape(-1,3).detach().cpu()), 0)
+            truth = torch.cat((truth, tgt[:,-x1*x2*x3:,:].flatten().detach().cpu()), 0)
+            test_result = torch.cat((test_result, output[:,-x1*x2*x3:,:].flatten().detach().cpu()), 0)
 
         a = torch.cat([test_ts.unsqueeze(-1), test_coord, test_result.unsqueeze(-1), truth.unsqueeze(-1)], dim=-1)
         a = a.numpy()
@@ -131,16 +128,22 @@ def predict_model(model, test_loader, epoch, input_type='patch', patch_size=2, \
         fig, ax = plt.subplots(nrows =1, ncols=1, figsize=(20,10))
         plot_start = int(len(final_result['prediction'])*plot_range[0])
         plot_end = int(len(final_result['prediction'])*plot_range[1])
-        ax.plot(final_result['prediction'][plot_start:plot_end],label='forecast')
-        ax.plot(final_result['truth'][plot_start:plot_end],label = 'truth')
-        ax.plot(final_result['prediction'][plot_start:plot_end]-final_result['truth'][plot_start:plot_end],ls='--',label='residual')
+        if scale==True:
+            ax.plot(scaler.inverse_transform(final_result['truth'][plot_start:plot_end]),label = 'truth')
+            ax.plot(scaler.inverse_transform(final_result['prediction'][plot_start:plot_end]),label='forecast')
+            ax.plot(scaler.inverse_transform(final_result['prediction'][plot_start:plot_end]) - \
+                    scaler.inverse_transform(final_result['truth'][plot_start:plot_end]),ls='--',label='residual')
+        elif scale==False:
+            ax.plot(final_result['truth'][plot_start:plot_end],label = 'truth')
+            ax.plot(final_result['prediction'][plot_start:plot_end],label='forecast')
+            ax.plot(final_result['prediction'][plot_start:plot_end] - final_result['truth'][plot_start:plot_end],ls='--',label='residual')            
         #ax.grid(True, which='both')
         ax.axhline(y=0)
         ax.legend(loc="upper right")
         if final_prediction == True:
             fig.savefig(root_dir + '/figs/transformer_pred.png')
         else:
-            fig.savefig(root_dir + f'/figs/epoch{epoch}_{pe_type}_{input_type}.png')
+            fig.savefig(root_dir + f'/figs/epoch{epoch}_{pe_type}_{window_size}_{x1}-{x2}-{x3}.png')
         plt.close(fig)
 
     if final_prediction == True:
@@ -154,15 +157,13 @@ if __name__ == "__main__":
     root_dir = '/scratch/zh2095/tune_results'
     sns.set_style("whitegrid")
     sns.set_palette(['#57068c','#E31212','#01AD86'])
-    best_config = {'input_type': 'patch', 'pe_type': '3d_temporal', 'patch_size': 2, 'feature_size': 768, 'num_enc_layers': 1, 'num_dec_layers': 1,\
-                     'num_head': 2, 'd_ff': 2048, 'dropout': 0.1, 'lr': 1e-5, 'lr_decay': 0.9, 'window_size': 10, 'batch_size': 256}
-    train_proportion = 0.6
-    test_proportion = 0.2
-    val_proportion = 0.2
-
+    best_config = {'window_size': 10, 'patch_size': (4,4,1), 'pe_type': '3d', 'batch_size': 1, 'feature_size': 120, 'num_enc_layers': 1, 'num_dec_layers': 1,\
+                     'num_head': 2, 'd_ff': 512, 'dropout': 0.1, 'lr': 1e-5, 'lr_decay': 0.9, 'scale': False}
+    # model hyperparameters
+    window_size = best_config['window_size']
     patch_size = best_config['patch_size']
-    input_type = best_config['input_type']
     pe_type = best_config['pe_type']
+    batch_size = best_config['batch_size']
     feature_size = best_config['feature_size']
     num_enc_layers = best_config['num_enc_layers']
     num_dec_layers = best_config['num_dec_layers']
@@ -171,11 +172,24 @@ if __name__ == "__main__":
     dropout = best_config['dropout']
     lr = best_config['lr']
     lr_decay = best_config['lr_decay']
-    window_size = best_config['window_size']
-    batch_size = best_config['batch_size']
+    scale = best_config['scale']
+
+    # dataset parameters
+    train_proportion = 0.6
+    test_proportion = 0.2
+    val_proportion = 0.2
+    grid_size = 16
+
+    # save model options
+    skip_training = False
+    save_model = True
+
+    print('-'*20 + ' Setting ' + '-'*20, flush=True)
+    print(f'pe type:{pe_type}, window size:{window_size}, patch size:{patch_size}', flush=True)
+    print('-'*50, flush=True)
 
     model = Tranformer(feature_size=feature_size,num_enc_layers=num_enc_layers,num_dec_layers = num_dec_layers,\
-            d_ff = d_ff, dropout=dropout,num_head=num_head,pe_type=pe_type)
+            d_ff = d_ff, dropout=dropout,num_head=num_head,pe_type=pe_type,grid_size=grid_size)
 
 
     device = "cpu"
@@ -188,6 +202,7 @@ if __name__ == "__main__":
                 
     criterion = nn.MSELoss()
     optimizer = optim.AdamW(model.parameters(), lr=lr)
+    #optimizer = optim.SGD(model.parameters(), lr=lr, momentum=0.9)
     scheduler = optim.lr_scheduler.StepLR(optimizer, step_size=1, gamma=lr_decay)
     writer = tensorboard.SummaryWriter('/scratch/zh2095/tensorboard_output/')
         
@@ -197,19 +212,19 @@ if __name__ == "__main__":
         #     model.load_state_dict(model_state)
         #     optimizer.load_state_dict(optimizer_state)
             
-    train_loader, test_loader, scaler = get_data_loaders(train_proportion, test_proportion, val_proportion,\
-            window_size = window_size, pred_size = 1, batch_size = batch_size, num_workers = 2, pin_memory = False,\
-            use_coords = True, use_time = True, test_mode = True, input_type = input_type, patch_size=patch_size)
 
-    epochs = 20
+    train_loader, test_loader, scaler = get_data_loaders(train_proportion, test_proportion, val_proportion,\
+        pred_size = 1, batch_size = batch_size, num_workers = 2, pin_memory = False, use_coords = True, use_time = True,\
+        test_mode = True, scale = scale, window_size = window_size, patch_size = patch_size)
+
+    epochs = 2
     train_losses = []
     test_losses = []
     tolerance = 10
     best_test_loss = float('inf')
     Early_Stopping = early_stopping(patience=5)
     counter_old = 0
-    skip_training = False
-
+    x1, x2, x3 = patch_size
     if os.path.exists(root_dir+'/best_model.pth') and skip_training:
         model.load_state_dict(torch.load(root_dir+'/best_model.pth'))
     else:
@@ -217,24 +232,20 @@ if __name__ == "__main__":
             model.train() 
             total_loss = 0.
             start_time = time.time()
-            if input_type == 'patch':
-                patch_size = patch_size
-            else:
-                patch_size = 1
 
             for i, ((src, tgt), (src_coord, tgt_coord), (src_ts, tgt_ts)) in enumerate(train_loader):
                 src, tgt, src_coord, tgt_coord, src_ts, tgt_ts = src.to(device), tgt.to(device), \
                                                                 src_coord.to(device), tgt_coord.to(device), src_ts.to(device), tgt_ts.to(device)
                 optimizer.zero_grad()
-                output, truth = process_one_batch(src, tgt, src_coord, tgt_coord, src_ts, tgt_ts)
-                loss = criterion(output[:,-patch_size**3:,:], tgt[:,-patch_size**3:,:])
+                output, truth = process_one_batch(src, tgt, src_coord, tgt_coord, src_ts, tgt_ts, patch_size)
+                loss = criterion(output[:,-x1*x2*x3:,:], tgt[:,-x1*x2*x3:,:])
                 total_loss += loss.item()
                 loss.backward()
                 optimizer.step()
 
-            avg_train_loss = total_loss*batch_size*patch_size**3/len(train_loader.dataset)
-            total_test_loss = evaluate(model, test_loader, criterion, input_type=input_type, patch_size=patch_size)
-            avg_test_loss = total_test_loss*patch_size**3/len(test_loader.dataset)
+            avg_train_loss = total_loss*batch_size/len(train_loader.dataset)
+            total_test_loss = evaluate(model, test_loader, criterion, patch_size=patch_size)
+            avg_test_loss = total_test_loss/len(test_loader.dataset)
 
             
             train_losses.append(avg_train_loss)
@@ -244,13 +255,12 @@ if __name__ == "__main__":
             if epoch==1: ###DEBUG
                 print(f'Total of {len(train_loader.dataset)} samples in training set and {len(test_loader.dataset)} samples in test set', flush=True)
 
-            print(f'Epoch: {epoch}, train_loss: {avg_train_loss}, test_loss: {avg_test_loss}, lr: {scheduler.get_last_lr()}', flush=True)
-            print(f'Training time for 1 epoch with pe type {pe_type} and input type {input_type}:  {time.time()-start_time} s', flush=True)
+            print(f'Epoch: {epoch}, train_loss: {avg_train_loss}, test_loss: {avg_test_loss}, lr: {scheduler.get_last_lr()}, training time: {time.time()-start_time} s', flush=True)
 
             if (epoch%2 == 0):
                 print(f'Saving prediction for epoch {epoch}', flush=True)
-                predict_model(model, test_loader, epoch, input_type=input_type, patch_size=patch_size, \
-                                    plot=True, plot_range=[0.5,0.51], final_prediction=False)   
+                predict_model(model, test_loader, epoch, patch_size=patch_size, scale=scale, \
+                                    plot=True, plot_range=[0,0.01], final_prediction=False)   
 
             writer.add_scalar('train_loss',avg_train_loss,epoch)
             writer.add_scalar('test_loss',avg_test_loss,epoch)
@@ -264,29 +274,28 @@ if __name__ == "__main__":
             if Early_Stopping.early_stop:
                 break
 
-
-
         #save model
-        if os.path.exists(root_dir + '/best_model.pth'):  # checking if there is a file with this name
-            os.remove(root_dir + '/best_model.pth')  # deleting the file
-            torch.save(model.state_dict(), root_dir + '/best_model.pth')
-        else:
-            torch.save(model.state_dict(), root_dir + '/best_model.pth')
+        if save_model:
+            if os.path.exists(root_dir + '/best_model.pth'):  # checking if there is a file with this name
+                os.remove(root_dir + '/best_model.pth')  # deleting the file
+                torch.save(model.state_dict(), root_dir + '/best_model.pth')
+            else:
+                torch.save(model.state_dict(), root_dir + '/best_model.pth')
 
 ### Plot losses        
     xs = np.arange(len(train_losses))
     fig, ax = plt.subplots(nrows =1, ncols=1, figsize=(20,10))
     ax.plot(xs,train_losses)
-    fig.savefig(root_dir + f'/figs/train_loss_{pe_type}_{input_type}.png')
+    fig.savefig(root_dir + f'/figs/train_loss_{pe_type}_{window_size}_{x1}-{x2}-{x3}.png')
     plt.close(fig)
     fig, ax = plt.subplots(nrows =1, ncols=1, figsize=(20,10))
     ax.plot(xs,test_losses)
-    fig.savefig(root_dir + f'/figs/test_loss_{pe_type}_{input_type}.png')
+    fig.savefig(root_dir + f'/figs/test_loss_{pe_type}_{window_size}_{x1}-{x2}-{x3}.png')
     plt.close(fig)
 
 ### Predict
     start_time = time.time()
-    final_result = predict_model(model, test_loader,  epoch, input_type=input_type, patch_size=patch_size, \
+    final_result = predict_model(model, test_loader,  epoch, patch_size=patch_size, scale=scale, \
                                             plot=True, plot_range=[0,1], final_prediction=True)
     prediction = final_result['prediction']
     print('-'*20 + ' Measure for Simulation Speed ' + '-'*20)
@@ -304,6 +313,23 @@ if __name__ == "__main__":
     print('-'*20 + ' Dataframe for Final Result ' + '-'*20)
     print(test_result_df)
     test_result_df.to_csv(root_dir + '/transformer_prediction.csv')
+
+### slice plot
+    img_dir = root_dir + '/slice_plot' 
+    pred_df = pd.read_csv(root_dir + '/transformer_prediction.csv',index_col=0)
+    grid_size = [16,16,16]
+    axis_colnames = ['x1','x2','x3']
+    slice_axis_index = 0
+    pred_colname = 'prediction'
+    truth_colname = 'truth'
+    time_colname = 'time'
+    plot_forecast(pred_df=pred_df, grid_size=grid_size, axis_colnames=axis_colnames, slice_axis_index=2, \
+                        pred_colname=pred_colname,truth_colname=truth_colname, time_colname=time_colname,  \
+                        plot_anime = True, img_dir = img_dir)  
+
+
+
+
